refactor: optimize rpc app mesh module
