+++ capa-spi-aws-config/src/main/java/group/rxcloud/capa/spi/aws/config/package-info.java
@@ -14,5 +14,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package group.rxcloud.capa.spi.aws.config;
