+++ pom.xml
@@ -80,7 +80,7 @@
         <maven.resource.version>3.2.0</maven.resource.version>
         <maven.surefire.version>3.0.0-M1</maven.surefire.version>
         <maven.jacoco.version>0.8.6</maven.jacoco.version>
-        <revision>1.0.1-SNAPSHOT</revision>
+        <revision>1.0.1.RELEASE</revision>
     </properties>
 
     <dependencyManagement>
@@ -93,6 +93,11 @@
             </dependency>
 
             <!-- SDK modules -->
+            <dependency>
+                <groupId>group.rxcloud</groupId>
+                <artifactId>capa-spi-aws</artifactId>
+                <version>${revision}</version>
+            </dependency>
             <dependency>
                 <groupId>group.rxcloud</groupId>
                 <artifactId>capa-spi-aws-infrastructure</artifactId>
